+++ miden/src/tools/mod.rs
@@ -1,7 +1,10 @@
 use core::fmt;
-use std::path::PathBuf;
+use std::{path::PathBuf, sync::Arc};
 
-use assembly::diagnostics::{IntoDiagnostic, Report, WrapErr};
+use assembly::{
+    DefaultSourceManager, SourceManager,
+    diagnostics::{Report, WrapErr},
+};
 use clap::Parser;
 use miden_vm::{DefaultHost, Host, Operation, StackInputs, internal::InputFile};
 use processor::{AsmOpInfo, TraceLenSummary};
@@ -46,12 +49,14 @@ impl Analyze {
             .to_lowercase();
 
         // Use a single match expression to load the program.
-        let program = match ext.as_str() {
-            "masp" => get_masp_program(&self.program_file)?,
+        let (program, source_manager) = match ext.as_str() {
+            "masp" => (
+                get_masp_program(&self.program_file)?,
+                Arc::new(DefaultSourceManager::default()) as Arc<dyn SourceManager>,
+            ),
             "masm" => get_masm_program(&self.program_file, &libraries)?,
             _ => return Err(Report::msg("The provided file must have a .masm or .masp extension")),
         };
-        // let program_hash: [u8; 32] = program.hash().into();
 
         // load input data from file
         let input_data = InputFile::read(&self.input_file, &self.program_file)?;
@@ -59,11 +64,11 @@ impl Analyze {
         // fetch the stack and program inputs from the arguments
         let stack_inputs = input_data.parse_stack_inputs().map_err(Report::msg)?;
         let mut host = DefaultHost::new(input_data.parse_advice_provider().map_err(Report::msg)?);
-        host.load_mast_forest(StdLibrary::default().mast_forest().clone())
-            .into_diagnostic()?;
+        host.load_mast_forest(StdLibrary::default().mast_forest().clone())?;
 
         let execution_details: ExecutionDetails =
-            analyze(&program, stack_inputs, host).expect("Could not retrieve execution details");
+            analyze(&program, stack_inputs, host, source_manager)
+                .expect("Could not retrieve execution details");
         let program_name = self
             .program_file
             .file_name()
@@ -235,17 +240,19 @@ fn analyze<H>(
     program: &Program,
     stack_inputs: StackInputs,
     mut host: H,
+    source_manager: Arc<dyn SourceManager>,
 ) -> Result<ExecutionDetails, Report>
 where
     H: Host,
 {
     let mut execution_details = ExecutionDetails::default();
 
-    let vm_state_iterator = processor::execute_iter(program, stack_inputs, &mut host);
+    let vm_state_iterator =
+        processor::execute_iter(program, stack_inputs, &mut host, source_manager);
     execution_details.set_trace_len_summary(vm_state_iterator.trace_len_summary());
 
     for state in vm_state_iterator {
-        let vm_state = state.into_diagnostic().wrap_err("execution error")?;
+        let vm_state = state.wrap_err("execution error")?;
         if matches!(vm_state.op, Some(Operation::Noop)) {
             execution_details.incr_noop_count();
         }
@@ -309,10 +316,11 @@ impl AsmOpStats {
 
 #[cfg(test)]
 mod tests {
+    use assembly::DefaultSourceManager;
     use miden_vm::Assembler;
     use processor::{ChipletsLengths, DefaultHost, TraceLenSummary};
 
-    use super::{AsmOpStats, ExecutionDetails, StackInputs};
+    use super::{AsmOpStats, ExecutionDetails, StackInputs, *};
 
     #[test]
     fn analyze_test() {
@@ -321,7 +329,8 @@ mod tests {
         let host = DefaultHost::default();
         let program = Assembler::default().with_debug_mode(true).assemble_program(source).unwrap();
         let execution_details =
-            super::analyze(&program, stack_inputs, host).expect("analyze_test: Unexpected Error");
+            super::analyze(&program, stack_inputs, host, Arc::new(DefaultSourceManager::default()))
+                .expect("analyze_test: Unexpected Error");
         let expected_details = ExecutionDetails {
             total_noops: 2,
             asm_op_stats: vec![
@@ -347,7 +356,8 @@ mod tests {
         let stack_inputs = vec![1, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0];
         let stack_inputs = StackInputs::try_from_ints(stack_inputs).unwrap();
         let host = DefaultHost::default();
-        let execution_details = super::analyze(&program, stack_inputs, host);
+        let execution_details =
+            super::analyze(&program, stack_inputs, host, Arc::new(DefaultSourceManager::default()));
         let expected_error = "Execution Error: DivideByZero(1)";
         assert_eq!(execution_details.err().unwrap().to_string(), expected_error);
     }
@@ -358,7 +368,8 @@ mod tests {
         let program = Assembler::default().with_debug_mode(true).assemble_program(source).unwrap();
         let stack_inputs = StackInputs::default();
         let host = DefaultHost::default();
-        let execution_details = super::analyze(&program, stack_inputs, host);
+        let execution_details =
+            super::analyze(&program, stack_inputs, host, Arc::new(DefaultSourceManager::default()));
         let expected_error = "Assembly Error: ParsingError(\"unexpected token: expected 'begin' but was 'mem_storew.1'\")";
         assert_eq!(execution_details.err().unwrap().to_string(), expected_error);
     }