+++ processor/src/chiplets/memory/tests.rs
@@ -1,5 +1,6 @@
 use alloc::vec::Vec;
 
+use assembly::SourceSpan;
 use miden_air::{
     RowIndex,
     trace::chiplets::memory::{
@@ -16,7 +17,7 @@ use super::{
     FieldElement, Memory, ONE, TraceFragment, V_COL_RANGE, WORD_COL_IDX,
     segment::{MemoryAccessType, MemoryOperation},
 };
-use crate::{ContextId, ExecutionError};
+use crate::{ContextId, MemoryError, errors::ErrorContext};
 
 #[test]
 fn mem_init() {
@@ -27,31 +28,32 @@ fn mem_init() {
 
 #[test]
 fn mem_read() {
+    let error_context = ErrorContext::default();
     let mut mem = Memory::default();
 
     // read a value from address 0; clk = 1
     let addr0 = ZERO;
-    let value = mem.read(ContextId::root(), addr0, 1.into()).unwrap();
+    let value = mem.read(ContextId::root(), addr0, 1.into(), &error_context).unwrap();
     assert_eq!(ZERO, value);
     assert_eq!(1, mem.num_accessed_words());
     assert_eq!(1, mem.trace_len());
 
     // read a value from address 3; clk = 2
     let addr3 = Felt::from(3_u32);
-    let value = mem.read(ContextId::root(), addr3, 2.into()).unwrap();
+    let value = mem.read(ContextId::root(), addr3, 2.into(), &error_context).unwrap();
     assert_eq!(ZERO, value);
     assert_eq!(1, mem.num_accessed_words());
     assert_eq!(2, mem.trace_len());
 
     // read a value from address 0 again; clk = 3
-    let value = mem.read(ContextId::root(), addr0, 3.into()).unwrap();
+    let value = mem.read(ContextId::root(), addr0, 3.into(), &error_context).unwrap();
     assert_eq!(ZERO, value);
     assert_eq!(1, mem.num_accessed_words());
     assert_eq!(3, mem.trace_len());
 
     // read a value from address 2; clk = 4
     let addr2 = Felt::from(2_u32);
-    let value = mem.read(ContextId::root(), addr2, 4.into()).unwrap();
+    let value = mem.read(ContextId::root(), addr2, 4.into(), &error_context).unwrap();
     assert_eq!(ZERO, value);
     assert_eq!(1, mem.num_accessed_words());
     assert_eq!(4, mem.trace_len());
@@ -116,22 +118,30 @@ fn mem_read_word_unaligned() {
     let addr = ONE;
     let clk = 1.into();
     let ctx = ContextId::root();
-    let ret = mem.read_word(ctx, addr, clk);
+    let ret = mem.read_word(ctx, addr, clk, &ErrorContext::default());
 
     assert_matches!(
         ret,
-        Err(ExecutionError::MemoryUnalignedWordAccess { addr: _, ctx: _, clk: _ })
+        Err(MemoryError::UnalignedWordAccess {
+            addr: _,
+            ctx: _,
+            clk: _,
+            label: SourceSpan::UNKNOWN,
+            source_file: None
+        })
     );
 }
 
 #[test]
 fn mem_write() {
+    let error_context = ErrorContext::default();
     let mut mem = Memory::default();
 
     // write a value into address 0; clk = 1
     let addr0 = 0_u32;
     let word1 = [ONE, ZERO, ZERO, ZERO];
-    mem.write_word(ContextId::root(), addr0.into(), 1.into(), word1).unwrap();
+    mem.write_word(ContextId::root(), addr0.into(), 1.into(), word1, &ErrorContext::default())
+        .unwrap();
     assert_eq!(word1, mem.get_word(ContextId::root(), addr0).unwrap().unwrap());
     assert_eq!(1, mem.num_accessed_words());
     assert_eq!(1, mem.trace_len());
@@ -139,7 +149,8 @@ fn mem_write() {
     // write a value into address 2; clk = 2
     let addr2 = 2_u32;
     let value5 = Felt::new(5);
-    mem.write(ContextId::root(), addr2.into(), 2.into(), value5).unwrap();
+    mem.write(ContextId::root(), addr2.into(), 2.into(), value5, &error_context)
+        .unwrap();
     assert_eq!(value5, mem.get_value(ContextId::root(), addr2).unwrap());
     assert_eq!(1, mem.num_accessed_words());
     assert_eq!(2, mem.trace_len());
@@ -147,7 +158,8 @@ fn mem_write() {
     // write a value into address 1; clk = 3
     let addr1 = 1_u32;
     let value7 = Felt::new(7);
-    mem.write(ContextId::root(), addr1.into(), 3.into(), value7).unwrap();
+    mem.write(ContextId::root(), addr1.into(), 3.into(), value7, &error_context)
+        .unwrap();
     assert_eq!(value7, mem.get_value(ContextId::root(), addr1).unwrap());
     assert_eq!(1, mem.num_accessed_words());
     assert_eq!(3, mem.trace_len());
@@ -155,7 +167,8 @@ fn mem_write() {
     // write a value into address 3; clk = 4
     let addr3 = 3_u32;
     let value9 = Felt::new(9);
-    mem.write(ContextId::root(), addr3.into(), 4.into(), value9).unwrap();
+    mem.write(ContextId::root(), addr3.into(), 4.into(), value9, &error_context)
+        .unwrap();
     assert_eq!(value9, mem.get_value(ContextId::root(), addr3).unwrap());
     assert_eq!(1, mem.num_accessed_words());
     assert_eq!(4, mem.trace_len());
@@ -163,14 +176,16 @@ fn mem_write() {
     // write a word into address 4; clk = 5
     let addr4 = 4_u32;
     let word1234 = [ONE, 2_u32.into(), 3_u32.into(), 4_u32.into()];
-    mem.write_word(ContextId::root(), addr4.into(), 5.into(), word1234).unwrap();
+    mem.write_word(ContextId::root(), addr4.into(), 5.into(), word1234, &ErrorContext::default())
+        .unwrap();
     assert_eq!(word1234, mem.get_word(ContextId::root(), addr4).unwrap().unwrap());
     assert_eq!(2, mem.num_accessed_words());
     assert_eq!(5, mem.trace_len());
 
     // write a word into address 0; clk = 6
     let word5678: [Felt; 4] = [5_u32.into(), 6_u32.into(), 7_u32.into(), 8_u32.into()];
-    mem.write_word(ContextId::root(), addr0.into(), 6.into(), word5678).unwrap();
+    mem.write_word(ContextId::root(), addr0.into(), 6.into(), word5678, &ErrorContext::default())
+        .unwrap();
     assert_eq!(word5678, mem.get_word(ContextId::root(), addr0).unwrap().unwrap());
     assert_eq!(2, mem.num_accessed_words());
     assert_eq!(6, mem.trace_len());
@@ -254,55 +269,63 @@ fn mem_write_word_unaligned() {
     let word1 = [ONE, ZERO, ZERO, ZERO];
     let clk = 1.into();
     let ctx = ContextId::root();
-    let ret = mem.write_word(ctx, addr, clk, word1);
+    let ret = mem.write_word(ctx, addr, clk, word1, &ErrorContext::default());
 
     assert_matches!(
         ret,
-        Err(ExecutionError::MemoryUnalignedWordAccess { addr: _, ctx: _, clk: _ })
+        Err(MemoryError::UnalignedWordAccess {
+            addr: _,
+            ctx: _,
+            clk: _,
+            label: SourceSpan::UNKNOWN,
+            source_file: None
+        })
     );
 }
 
 /// Tests that values written are properly read back.
 #[test]
 fn mem_write_read() {
+    let error_context = ErrorContext::default();
     let mut mem = Memory::default();
     let mut clk: RowIndex = 1.into();
 
     // write [1,2,3,4] starting at address 0; clk = 1
     let word1234 = [ONE, 2_u32.into(), 3_u32.into(), 4_u32.into()];
-    mem.write_word(ContextId::root(), ZERO, clk, word1234).unwrap();
+    mem.write_word(ContextId::root(), ZERO, clk, word1234, &error_context).unwrap();
     clk += 1_u32;
 
     // read individual values from addresses 3,2,1,0; clk = 2,3,4,5
-    let value_read = mem.read(ContextId::root(), 3_u32.into(), clk).unwrap();
+    let value_read = mem.read(ContextId::root(), 3_u32.into(), clk, &error_context).unwrap();
     assert_eq!(value_read, 4_u32.into());
     clk += 1_u32;
-    let value_read = mem.read(ContextId::root(), 2_u32.into(), clk).unwrap();
+    let value_read = mem.read(ContextId::root(), 2_u32.into(), clk, &error_context).unwrap();
     assert_eq!(value_read, 3_u32.into());
     clk += 1_u32;
-    let value_read = mem.read(ContextId::root(), 1_u32.into(), clk).unwrap();
+    let value_read = mem.read(ContextId::root(), 1_u32.into(), clk, &error_context).unwrap();
     assert_eq!(value_read, 2_u32.into());
     clk += 1_u32;
-    let value_read = mem.read(ContextId::root(), ZERO, clk).unwrap();
+    let value_read = mem.read(ContextId::root(), ZERO, clk, &error_context).unwrap();
     assert_eq!(value_read, 1_u32.into());
     clk += 1_u32;
 
     // read word from address 0; clk = 6
-    let word_read = mem.read_word(ContextId::root(), ZERO, clk).unwrap();
+    let word_read = mem.read_word(ContextId::root(), ZERO, clk, &error_context).unwrap();
     assert_eq!(word_read, word1234);
     clk += 1_u32;
 
     // write 42 into address 2; clk = 7
-    mem.write(ContextId::root(), 2_u32.into(), clk, 42_u32.into()).unwrap();
+    mem.write(ContextId::root(), 2_u32.into(), clk, 42_u32.into(), &error_context)
+        .unwrap();
     clk += 1_u32;
 
     // read element from address 2; clk = 8
-    let value_read = mem.read(ContextId::root(), 2_u32.into(), clk).unwrap();
+    let value_read = mem.read(ContextId::root(), 2_u32.into(), clk, &error_context).unwrap();
     assert_eq!(value_read, 42_u32.into());
     clk += 1_u32;
 
     // read word from address 0; clk = 9
-    let word_read = mem.read_word(ContextId::root(), ZERO, clk).unwrap();
+    let word_read = mem.read_word(ContextId::root(), ZERO, clk, &error_context).unwrap();
     assert_eq!(word_read, [ONE, 2_u32.into(), 42_u32.into(), 4_u32.into()]);
     clk += 1_u32;
 
@@ -421,12 +444,24 @@ fn mem_get_state_at() {
     // Write word starting at (ctx = 0, addr = 40) at clk = 1.
     // This means that mem[40..43] is set at the beginning of clk = 2
     let word1234 = [ONE, 2_u32.into(), 3_u32.into(), 4_u32.into()];
-    mem.write_word(ContextId::root(), addr_start.into(), 1.into(), word1234)
-        .unwrap();
+    mem.write_word(
+        ContextId::root(),
+        addr_start.into(),
+        1.into(),
+        word1234,
+        &ErrorContext::default(),
+    )
+    .unwrap();
 
     let word4567: [Felt; 4] = [4_u32.into(), 5_u32.into(), 6_u32.into(), 7_u32.into()];
-    mem.write_word(ContextId::root(), addr_start.into(), 2.into(), word4567)
-        .unwrap();
+    mem.write_word(
+        ContextId::root(),
+        addr_start.into(),
+        2.into(),
+        word4567,
+        &ErrorContext::default(),
+    )
+    .unwrap();
 
     // Check memory state at clk = 2
     let clk: RowIndex = 2.into();